Pass on provided initialization args even if all mutexes are null. (#50)
The null check is not needed here as since #5 the
JNA_CK_C_INITIALIZE_ARGS handles those.
And the big issue with the old code is that if the args are not
passed on any provided flag is also not. This could cause Java
process crashes as CKF_OS_LOCKING_OK is missing.