Use try-finally so FindObjectsFinal always gets called (#54)
It is possible that the inner FindObjects call fails in the single-part
FindObjects method. In that case it is necessary to call FindObjectsFinal,
otherwise the session enters an inconsistent state, and later Cryptoki
calls fail with CKR.OPERATION_ACTIVE.

This commit wraps the code in try-finally so FindObjectsFinal gets called
also if there is an error.